Move extensions with dependencies to extensions.contrib. Move extension without dependency to extensions. Make extensions discoverable by markdown with dot import notation. Add dependencies for extensions to extras_require in setup.py. Rename markup tags to markymark. Update docs on extensions and template tag usage.
