+++ docs/source/modules/markymark.extensions.contrib.rst
@@ -1,0 +1,30 @@
+markymark.extensions.contrib package
+====================================
+
+Submodules
+----------
+
+markymark.extensions.contrib.anylink module
+-------------------------------------------
+
+.. automodule:: markymark.extensions.contrib.anylink
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+markymark.extensions.contrib.filer module
+-----------------------------------------
+
+.. automodule:: markymark.extensions.contrib.filer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: markymark.extensions.contrib
+    :members:
+    :undoc-members:
+    :show-inheritance:
