+++ docs/source/modules/markymark.extensions.rst
@@ -1,0 +1,29 @@
+markymark.extensions package
+============================
+
+Subpackages
+-----------
+
+.. toctree::
+
+    markymark.extensions.contrib
+
+Submodules
+----------
+
+markymark.extensions.autolink module
+------------------------------------
+
+.. automodule:: markymark.extensions.autolink
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+Module contents
+---------------
+
+.. automodule:: markymark.extensions
+    :members:
+    :undoc-members:
+    :show-inheritance:
