+++ docs/source/modules/markymark.rst
@@ -6,6 +6,7 @@ Subpackages
 
 .. toctree::
 
+    markymark.extensions
     markymark.templatetags
 
 Submodules
@@ -15,14 +16,6 @@ markymark.conf module
 ---------------------
 
 .. automodule:: markymark.conf
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-markymark.extensions module
----------------------------
-
-.. automodule:: markymark.extensions
     :members:
     :undoc-members:
     :show-inheritance:
