+++ markymark/templatetags/markymark.py
@@ -1,3 +1,5 @@
+from __future__ import absolute_import
+
 from django import template
 from django.utils.safestring import mark_safe
